Added __main__.py so the code can now be called with python -m src x. Mostly deleted extraneous code from multi_processing.py (some remains to port over to single_processing.py), and moved multiprocessing methods from __main__.py to multi_processing.py
